Began implementation of rewriting engine
Reduce.jl seems to be more trouble than its worth. Here is the begging
of an engine for processing rewriting rules for julia expressions.

There are a few worts with julia expressions which we try to ignore.
One is that julia has / and //. We use / for all purposes and hope we
don't see a //.

The rewriting engine is mainly about matching up variables in a match
expression (i.e. the first half of a rule) against arbitrary
expressions. Most functions are not treated specially but we give
special treatment to expressions invloving +, *, -, etc in the match
expression. In particular, we try to convert the matcher and the thing
we are matching against into a word in the corresponding group (that
is, a sequence of expressions and powers) and then match in a sensible
way based on that.

This method isn't perfect. There's no way to specify that some symbol
may take the identity element (so we assume that it cannot) and we
currently don't consider splitting up powers as part of the matching,
for example to match a+b against 5+5, 5+5 gets converted into the word
2*5 and then we can't split this into 1*5+1*5 to get the right match.

Also a+a can succesfully match against 5+5 and 2*5 as 2*5 gets
converted into the word (1*(2*5)) in the group of +, and then a*2 is
recursively matched against 2*5 (i.e. in the group of *) which is
successful. Whereas a*2 cannot match against 5+5.

The main entry point is Rewrite.rewrite.

The rules themselves have not yet been written.