More work on rewriting.
Fixed some bugs
Changed expansion to work better by implementing as a function insted.
Added some canonicalisation to the main rewriter
Wrote some rules for canonicalise but they are much to slow for
something which is supposed to not do much. I think I'm going to need
to do some kind of ring-sensitive stuff in the canonicalise in
Rewrite.jl. Maybe change products so they canonicalise with any
integer factors in the second argument of an outer *(). Then they
might be picked up when canonicalising sums. Also we should probably
change the way sums and products are rewritten so that e.g.
a::Integer+b::Integer can match against 1 + 2 + x.

It may be worthwhile to completely change matching to work with
coroutines so that we don't need to find all matches. It seems that
sometimes it is usefull to know all matches though.