Update to new wrui. Partially update to julia v0.6
Annoyingly these two changes happened to coincide.

Changing to julia v0.6 seems to have broken some of the rewriting code
and I've not managed to fix it all yet. In fact I think I might just
need to throw a lot of it out anyway. It turns out Julia's expression
format is awful so a v2 plan for Rewrite is to change from using Expr
to just using S-expressions (like almost every sane implementation of
rewriting). I still quite like the idea of matching based on some
algebraic properties of +, * but I'm starting to think that actually
it doesn't work as well as one would hope so perhaps I will get rid of
it. If I do that then I would need to decide how to deal with +, -,
etc and I think that could be done by forcing them all to be
binary (or unary too for -) so that any term/sum can be split into
precisely two things. Then we would need to do some kind of processing
to layout. Alternatively, we could keep the operator applications as
nary but let them be treated specially as binary in the matcher so
that they matchers can be simple.

The changes to v0.6 are not totally finished. Julia still throws up
lots of warnings.

The change to the new wrui should allow for keyboard input which is nice.